+++ packages/event-listener/src/adaptors/neo4j/index.ts
@@ -1,0 +1,13 @@
+import {
+  contentPropertySet,
+  linkPropertySet,
+  contentLinked,
+  contentCreated
+} from "./eventHandlers";
+
+export default {
+  contentPropertySet,
+  linkPropertySet,
+  contentLinked,
+  contentCreated
+};
